+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Remove random function from ink!.
+
 ## Version 4.0.0-alpha.3
 
 ### Breaking Changes